Fix: nav_graph all fragments, Long types, Safe Args references
- nav_graph: dodano categoryDetailFragment i calendarFragment z akcjami
- mementoId zostaje jako Long wszedie (bez konwersji Int)
- Poprawiono porownania 0L, toInt() dla AlarmScheduler
- CategoriesFragmentDirections/CategoryDetailFragmentArgs teraz generowane

+++ app/src/main/java/com/pawel/memento/ui/categories/CategoryDetailFragment.kt
@@ -30,9 +30,7 @@ class CategoryDetailFragment : Fragment() {
             onToggleComplete = { viewModel.toggleCompleted(it.memento) },
             onClick = {
                 findNavController().navigate(
-                    CategoryDetailFragmentDirections.actionCategoryDetailFragmentToMementoEditorFragment(
-                        it.memento.id.toInt()
-                    )
+                    CategoryDetailFragmentDirections.actionCategoryDetailFragmentToMementoEditorFragment(it.memento.id)
                 )
             },
             onLongClick = { item ->
@@ -47,7 +45,7 @@ class CategoryDetailFragment : Fragment() {
         binding.recyclerView.adapter = adapter
         binding.fabAdd.setOnClickListener {
             findNavController().navigate(
-                CategoryDetailFragmentDirections.actionCategoryDetailFragmentToMementoEditorFragment(0)
+                CategoryDetailFragmentDirections.actionCategoryDetailFragmentToMementoEditorFragment(0L)
             )
         }
         viewModel.setSelectedCategory(args.categoryId)
